+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { CartService } from 'src/app/cart/services/cart.service';
 import { Product } from '../../models/product.model';
 import { ProductsService } from '../../services/products.service';
 
@@ -8,15 +9,18 @@ import { ProductsService } from '../../services/products.service';
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
-
   @Input() products: Array<Product>;
 
-  // в конструкторе только внедрение зависимости
-  constructor(private productsService: ProductsService) {
-  }
+  constructor(
+    private productsService: ProductsService,
+    private cartService: CartService
+  ) {}
 
   ngOnInit(): void {
     this.products = this.productsService.getProducts();
   }
 
+  onBuyProduct(product: Product): void {
+    this.cartService.addToCart(product);
+  }
 }