+++ src/app/cart/models/cart-item.model.ts
@@ -3,8 +3,6 @@ import { Product } from '../../products/models/product.model';
 export class CartItem {
   constructor(
     public product: Product,
-    public count: number = 1,
-    public price: number = product.price,
-    public totalPrice: number = product.price
+    public count: number = 1
   ) { }
 }