+++ src/app/products/models/product.model.ts
@@ -6,6 +6,7 @@ export enum Categoty {
 
 export class Product {
   constructor(
+    public id: string,
     public name: string,
     public description: string,
     public price: number,